resolve conflict

+++ file.js
@@ -4,6 +4,5 @@ function abc(){
     a+blur;
 }
 
-function add (float a, float b){
-    
+function add (int a ,int b){
 }
